Add tests for newly added BitUtil functions

+++ huffman-compression/src/bit_util.cpp
@@ -21,6 +21,6 @@ int toByteIndex(int bitIndex) {
   return bitIndex / CHAR_BIT;
 }
 int toByteSize(int bitSize) {
-  return std::ceil(bitSize / CHAR_BIT);
+  return std::ceil(bitSize / (double)CHAR_BIT);
 }
 }
